refactor(resource): extract resource selectors and dialogs into features/resource
The Agent/Assistant/Workspace selectors lived in global components/Selector but
lazy-loaded the library create/edit dialogs and imported pages/library types, so
a global component reverse-depended on a page (#16187 review point 1).
PromptEditDialog sat in the components/ root though it edits library prompt
resources (point 3).

Gather the resource domain into src/renderer/features/resource/ per
naming-conventions.md §4.10:
- selectors: Agent/Assistant/Workspace + ResourceSelectorShell
- dialogs: create/edit dialogs (from pages/library/dialogs) + PromptEditDialog

components/Selector keeps only the generic model/shell/Basic primitives; the
reusable PromptEditorField stays in components/. adapters/constants/types remain
in pages/library and are imported by the feature.
