+++ src/renderer/components/Selector/index.tsx
@@ -1,3 +1,2 @@
 export { default } from './BasicSelector'
 export * from './model'
-export * from './resource'
