+++ src/renderer/features/resource/AgentSelector.tsx
@@ -1,22 +1,22 @@
 import { loggerService } from '@logger'
+import type { SelectorShellMountStrategy, SelectorShellProps } from '@renderer/components/Selector/shell/SelectorShell'
 import { useMutation, useQuery } from '@renderer/data/hooks/useDataApi'
-import { useAgentModelFilter } from '@renderer/hooks/agents/useAgentModelFilter'
-import { usePins } from '@renderer/hooks/usePins'
 import {
   ResourceCreateDialog,
   type ResourceCreateDialogValues
-} from '@renderer/pages/library/dialogs/ResourceCreateDialog'
+} from '@renderer/features/resource/dialogs/ResourceCreateDialog'
+import { useAgentModelFilter } from '@renderer/hooks/agents/useAgentModelFilter'
+import { usePins } from '@renderer/hooks/usePins'
 import type { AgentDetail } from '@renderer/pages/library/types'
 import { getAgentAvatarFromConfiguration } from '@renderer/utils/agent'
 import { lazy, type ReactElement, Suspense, useCallback, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import type { SelectorShellMountStrategy, SelectorShellProps } from '../shell/SelectorShell'
 import { ResourceSelectorShell, type ResourceSelectorShellItem } from './ResourceSelectorShell'
 
 const logger = loggerService.withContext('AgentSelector')
 const AgentEditDialog = lazy(() =>
-  import('@renderer/pages/library/dialogs/edit/AgentEditDialog').then((module) => ({
+  import('@renderer/features/resource/dialogs/edit/AgentEditDialog').then((module) => ({
     default: module.AgentEditDialog
   }))
 )
