+++ src/renderer/features/resource/dialogs/components/CatalogPicker.tsx
