+++ src/renderer/features/resource/dialogs/components/DialogFormFields.tsx
