+++ src/renderer/features/resource/dialogs/components/McpServerAvatar.tsx
