+++ src/renderer/features/resource/dialogs/components/McpServerCatalogGrid.tsx
