+++ src/renderer/features/resource/dialogs/edit/AssistantEditDialog.tsx
