+++ src/renderer/features/resource/dialogs/edit/EditDialogShared.tsx
