+++ src/renderer/features/resource/dialogs/form/__tests__/agent.test.ts
@@ -1,6 +1,6 @@
+import type { AgentDetail } from '@renderer/pages/library/types'
 import { describe, expect, it } from 'vitest'
 
-import type { AgentDetail } from '../../../types'
 import {
   type AgentFormState,
   applyAgentFormPatch,
