+++ src/renderer/features/resource/dialogs/form/__tests__/assistant.test.ts
