+++ src/renderer/features/resource/dialogs/form/agent.ts
@@ -3,11 +3,10 @@ import {
   DEFAULT_HEARTBEAT_INTERVAL,
   normalizePermissionMode
 } from '@renderer/hooks/agents/permissionMode'
+import type { AgentDetail } from '@renderer/pages/library/types'
 import type { UpdateAgentDto } from '@shared/data/api/schemas/agents'
 import type { AgentConfiguration } from '@shared/data/types/agent'
 import type { UniqueModelId } from '@shared/data/types/model'
-
-import type { AgentDetail } from '../../types'
 
 // ---------------------------------------------------------------------------
 // Form state
