+++ src/renderer/features/resource/dialogs/form/assistant.ts
