+++ src/renderer/features/resource/dialogs/form/assistantModelFilter.ts
