+++ src/renderer/features/resource/dialogs/index.ts
